requirements.txt
pandas
PyYAML
tqdm
langchain
langchain-community
langchain-openai
torch
transformers
datasets
accelerate
vllm
requests